feat(search): Extract AmazingImageViewFactory to a separate file
