+++ AmazingApp/AmazingApp/Views/AmazingImageView/AmazingImageView.swift
@@ -7,20 +7,6 @@
 
 import UIKit
 
-struct AmazingImageViewBuilder {
-    static func makeView(frame: CGRect = .zero) -> AmazingImageView {
-        let presenter = AmazingImageViewPresenter()
-        let interactor = AmazingImageViewInteractor(presenter: presenter)
-        let view = AmazingImageView(
-            frame: frame,
-            interactor: interactor
-        )
-        presenter.view = view
-        
-        return view
-    }
-}
-
 final class AmazingImageView: UIImageView {
     
     private let activityIndicator: UIActivityIndicatorView = {
