--- AmazingApp/AmazingApp/Views/AmazingImageView/AmazingImageViewBuilder.swift
@@ -1,8 +1,0 @@
-//
-//  AmazingImageViewBuilder.swift
-//  AmazingApp
-//
-//  Created by Alysson on 01/08/22.
-//
-
-import Foundation
